revert renaming for_manager for_host
